Refactored seed.js and cryptoHelper.js file locations
